pyproject.toml
[project]
name = "obsidian-kg"
version = "0.1.0"
description = "Knowledge Graph AI Work Assistant - Build temporal knowledge graphs from Obsidian Vault"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "graphiti-core>=0.5",
    "neo4j>=5.26",
    "python-frontmatter>=1.1",
    "watchdog>=4.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
kg = "obsidian_kg.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_kg"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
